tests: FunctionalTests -> AudioExportTest & drop mute group test
the mute group test is disabled for as long as I am involved in this project. I drop it and replace it with a new one in case we encounter a bug/regression within the mute group feature
