+++ src/tests/registeredTests.h
@@ -26,6 +26,7 @@
 #include "AssertionTest.cpp"
 #include "AudioBenchmark.h"
 #include "AudioDriverTest.h"
+#include "AudioExportTest.h"
 #include "AutomationPathSerializerTest.cpp"
 #include "AutomationPathTest.cpp"
 #include "CliTest.h"
@@ -33,7 +34,6 @@
 #include "EventQueueTest.cpp"
 #include "DrumkitExportTest.h"
 #include "FilesystemTest.h"
-#include "FunctionalTests.cpp"
 #include "InstrumentListTest.cpp"
 #include "LicenseTest.h"
 #include "MemoryLeakageTest.h"
@@ -56,6 +56,7 @@ CPPUNIT_TEST_SUITE_REGISTRATION( ADSRTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( AssertionTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( AudioBenchmark );
 CPPUNIT_TEST_SUITE_REGISTRATION( AudioDriverTest );
+CPPUNIT_TEST_SUITE_REGISTRATION( AudioExportTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( AutomationPathSerializerTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( AutomationPathTest );
 #if not defined(WIN32) and not defined (__APPLE__)
@@ -66,7 +67,6 @@ CPPUNIT_TEST_SUITE_REGISTRATION( CoreActionControllerTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( EventQueueTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( DrumkitExportTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( FilesystemTest );
-CPPUNIT_TEST_SUITE_REGISTRATION( FunctionalTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( InstrumentListTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( LicenseTest );
 CPPUNIT_TEST_SUITE_REGISTRATION( MemoryLeakageTest );
